More features added in GUI (enable/disable button) Message shown after files are copied

+++ gui.py
@@ -2,7 +2,7 @@ from PyQt5 import QtCore, QtGui, QtWidgets
 from PyQt5.QtWidgets import QFileDialog, QLineEdit
 import os
 import shutil
-
+import time
 import sys
 
 from mainwindow import Ui_MainWindow  as window1 # Replace with the name of your generated UI file
@@ -35,6 +35,7 @@ class MainWindow(QtWidgets.QMainWindow, window1):
 
         # OK button
         self.ok.clicked.connect(self.on_ok_clicked)
+        self.ok.setEnabled(False)
         # Cancel button
         self.cancel.clicked.connect(self.on_cancel_clicked)
 
@@ -42,22 +43,26 @@ class MainWindow(QtWidgets.QMainWindow, window1):
         directory = QFileDialog.getExistingDirectory(self, "Select Directory")
         if directory:
             # Do something with the selected directory
-            print(f"Selected directory: {directory}")
+            # print(f"Selected directory: {directory}")
             self.srcpath.setText(str(directory))
     
     def on_button2_clicked(self):
         directory = QFileDialog.getExistingDirectory(self, "Select Directory")
         if directory:
             # Do something with the selected directory
-            print(f"Selected directory: {directory}")
+            # print(f"Selected directory: {directory}")
             self.destpath.setText(str(directory))
 
     def on_button3_clicked(self):
         filename, _ = QFileDialog.getOpenFileName(self, "Open File", "", "All Files (*);;Text Files (*.txt)")
         if filename:
             # Do something with the selected file
-            print(f"Selected file: {filename}")
+            # print(f"Selected file: {filename}")
             self.textfile.setText(str(filename))
+            src = self.srcpath.toPlainText()
+            dest = self.destpath.toPlainText()
+            if src!='' and dest!='':
+                self.ok.setEnabled(True)
     
     
     def on_ok_clicked(self):
@@ -66,6 +71,8 @@ class MainWindow(QtWidgets.QMainWindow, window1):
             src = self.srcpath.toPlainText()
             dest = self.destpath.toPlainText()
             textfile = self.textfile.toPlainText()
+            
+            self.successmsg.setText("Files Copying..........")
             
             if src=='' or dest=='' or textfile=='':
                 print("Enter Data First")
@@ -83,10 +90,18 @@ class MainWindow(QtWidgets.QMainWindow, window1):
                         source_path = os.path.join(src, filename)
                         destination_path = os.path.join(dest, filename)
 
+                        if os.path.exists(source_path):
+                            print("File exists")
+                            shutil.copy(source_path, destination_path)
+                        else:
+                            print("File does not exist")
                         # Move the file
-                        shutil.copy(source_path, destination_path)
                 # newWindow = AlertWindow()
                 # newWindow.show()
+                time.sleep(3)
+                self.successmsg.setText("Files Copied Successfully")
+                self.ok.setEnabled(False)
+                self.cancel.setText('Close')
             
     
     def on_cancel_clicked(self):
